requirements.txt
PyGithub>=2.4.0
PyYAML>=6.0
